requirements.txt
Django==1.6
django-crontab==0.6.0
django-smtp-ssl==1.0
MySQL-python==1.2.5
python-etcd==0.4.3
python-jenkins==0.4.13
python-ldap==2.4.22
South==1.0.2
uWSGI==2.0.14